Added / and renamed component
